ai/requirements.txt
google-generativeai>=0.8.0
mistralai>=1.0.0
python-dotenv==1.0.1
requests==2.31.0
flask==3.0.3
flask-cors==5.0.0